+++ crates/oxc_cli/src/command/lint.rs
@@ -1,4 +1,4 @@
-use std::path::PathBuf;
+use std::{path::PathBuf, str::FromStr};
 
 use bpaf::Bpaf;
 use oxc_linter::AllowWarnDeny;
@@ -136,8 +136,7 @@ pub struct WarningOptions {
 #[derive(Debug, Clone, Bpaf)]
 pub struct OutputOptions {
     /// Use a specific output format (default, json)
-    // last flag is the default
-    #[bpaf(long, short, flag(OutputFormat::Json, OutputFormat::Default))]
+    #[bpaf(long, short, fallback(OutputFormat::Default))]
     pub format: OutputFormat,
 }
 
@@ -145,6 +144,17 @@ pub struct OutputOptions {
 pub enum OutputFormat {
     Default,
     Json,
+}
+
+impl FromStr for OutputFormat {
+    type Err = String;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "json" => Ok(Self::Json),
+            "default" => Ok(Self::Default),
+            _ => Err(format!("'{s}' is not a known format")),
+        }
+    }
 }
 
 /// Enable Plugins
@@ -308,6 +318,16 @@ mod lint_options {
     fn format() {
         let options = get_lint_options("-f json");
         assert_eq!(options.output_options.format, OutputFormat::Json);
+        assert!(options.paths.is_empty());
+    }
+
+    #[test]
+    fn format_error() {
+        let args = "-f asdf".split(' ').map(std::string::ToString::to_string).collect::<Vec<_>>();
+        let result = lint_command().run_inner(args.as_slice());
+        assert!(result.is_err_and(
+            |err| err.unwrap_stderr() == "couldn't parse `asdf`: 'asdf' is not a known format"
+        ));
     }
 
     #[test]