+++ tasks/transform_conformance/src/test_case.rs
@@ -19,6 +19,7 @@ use oxc_transformer::{
 
 use crate::{fixture_root, root, TestRunnerEnv};
 
+#[derive(Debug)]
 pub enum TestCaseKind {
     Transform(ConformanceTestCase),
     Exec(ExecTestCase),
@@ -148,7 +149,7 @@ pub trait TestCase {
 
         let transformed_program = allocator.alloc(ret.program);
 
-        let result = Transformer::new(&allocator, source_type, semantic, self.transform_options())
+        let result = Transformer::new(&allocator, path, semantic, self.transform_options())
             .build(transformed_program);
 
         result.map(|()| {
@@ -159,6 +160,7 @@ pub trait TestCase {
     }
 }
 
+#[derive(Debug)]
 pub struct ConformanceTestCase {
     path: PathBuf,
     options: BabelOptions,
@@ -219,7 +221,7 @@ impl TestCase for ConformanceTestCase {
             .semantic;
         let program = allocator.alloc(ret.program);
         let transformer =
-            Transformer::new(&allocator, source_type, semantic, transform_options.clone());
+            Transformer::new(&allocator, &self.path, semantic, transform_options.clone());
 
         let codegen_options = CodegenOptions::default();
         let mut transformed_code = String::new();
@@ -284,6 +286,7 @@ impl TestCase for ConformanceTestCase {
     }
 }
 
+#[derive(Debug)]
 pub struct ExecTestCase {
     path: PathBuf,
     options: BabelOptions,