+++ crates/oxc_sourcemap/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this crate will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project does not adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) until v1.0.0.
 
+## [0.12.3] - 2024-04-11
+
+### Features
+
+- Add x_google_ignoreList (#2928)
+- Add sourceRoot (#2926)
+
 ## [0.12.2] - 2024-04-08
 
 ### Features