+++ crates/oxc_transformer/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this crate will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project does not adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) until v1.0.0.
 
+## [0.12.3] - 2024-04-11
+
+### Features
+
+- Implement plugin-transform-react-display-name top-down (#2937)
+- Add transform context to all plugins (#2931)
+- Add transform callback methods (#2929)
+- Add react preset (#2921)
+
 ## [0.11.1] - 2024-04-03
 
 ### Features